requirements.txt
customtkinter>=5.2.0
requests>=2.31.0
python-whois>=0.8.0
dnspython>=2.4.0
phonenumbers>=8.13.0
Pillow>=10.0.0
shodan>=1.31.0
beautifulsoup4>=4.12.0
PyMuPDF>=1.27.0
python-docx>=1.2.0
